Split the variant test into three translation units and add a shared header
